feat: enable deb otp

+++ backend/gateway/core/src/main/scala/io/mesazon/gateway/service/UserForgotPasswordService.scala
@@ -167,17 +167,11 @@ object UserForgotPasswordService {
               )
             )
           else if (
-            userOtpRow.otp != forgotPasswordVerifyOTP.otp || verifyOTPinDev(
-              userOtpRow.otp,
+            userOtpRow.otp == forgotPasswordVerifyOTP.otp || verifyOtpInDev(
+              forgotPasswordVerifyOTP.otp,
               isDev = userForgotPasswordConfig.isDev,
             )
           )
-            ZIO.fail(
-              ServiceError.BadRequestError.OtpVerifyError(
-                s"Wrong OTP provided for OTP ID [${forgotPasswordVerifyOTP.otpID}] and OTP type [${OtpType.ForgotPassword}]"
-              )
-            )
-          else
             userOtpRepository.deleteUserOtp(
               otpID = forgotPasswordVerifyOTP.otpID,
               userID = userOtpRow.userID,
@@ -188,6 +182,12 @@ object UserForgotPasswordService {
             ) *> userActionAttemptRepository.deleteUserActionAttempt(
               userID = userOtpRow.userID,
               actionAttemptType = ActionAttemptType.ForgotPasswordVerifyOTP,
+            )
+          else
+            ZIO.fail(
+              ServiceError.BadRequestError.OtpVerifyError(
+                s"Wrong OTP provided for OTP ID [${forgotPasswordVerifyOTP.otpID}] and OTP type [${OtpType.ForgotPassword}]"
+              )
             )
         resetPasswordJwt <- jwtService.generateResetPasswordToken(userDetailsRow.userID)
         _                <- userTokenRepository
